requirements.txt
Django==4.2.3
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
pillow==9.4.0
django-allauth==0.54.0
dj-rest-auth==4.0.1
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.2
